requirements.txt
attrs==22.1.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.3
django-environ==0.9.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-mptt-admin==2.3.0
djangorestframework==3.14.0
flake8==6.0.0
frozenlist==1.3.3
idna==3.4
isort==5.11.4
magic-filter==1.0.9
mccabe==0.7.0
multidict==6.0.3
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.6
sqlparse==0.4.3
yarl==1.8.2
